adicionado o btn info no map

+++ src/service/api.js
@@ -3,13 +3,18 @@ import axios from "axios";
 import { router } from "expo-router";
 
 const api = axios.create({
-    baseURL:'http://192.168.18.12:3007/api',
+    baseURL:'https://fileprecadastro.greenet.net.br//api',
     timeout: 5000, // 5 segundos
 })
 
 api.interceptors.response.use((response) => {
     return response;
 },(error) => {
+    console.log(JSON.stringify(error))
+    if(error.response && error.response.status === 404){
+        return Promise.reject({errors:["Algo deu errado tente novamente mais tarde"]});
+    }
+    
     if(error.response && error.response.status === 401){
         console.log("redirecionar")
         LoginBD.delete();
